chat-service/requirement.txt
boto3>=1.28.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
flask-restx>=1.3.0
flasgger>=0.9.7
Flask-SQLAlchemy>=3.0.0
psycopg2-binary>=2.9.6
PyPDF2>=3.0.0
python-docx
langchain~=0.1.20
langchain-core~=0.1.27
langchain-postgres~=0.0.3
tiktoken>=0.6.0
sentence-transformers~=2.7.0